Move classes to submodules
