+++ LEGO1/realtime/orientableroi.h
