+++ LEGO1/viewmanager/viewmanager.cpp
