+++ LEGO1/viewmanager/viewmanager.h
